Update Juni 2020

+++ app/Http/Controllers/PesertaController.php
@@ -10,7 +10,6 @@ use Illuminate\Support\Facades\Gate;
 
 class PesertaController extends Controller
 {
-    //
 
     public function main(Request $request){
 
@@ -18,12 +17,18 @@ class PesertaController extends Controller
             $query->where('nama', 'LIKE' , "%{$request->search}%");
         })->when($request->delegasi, function($query) use ($request) {
             $query->where('delegasi', "{$request->delegasi}");
-        })->paginate(8);
-        $pesertas->appends($request->only('search', 'delegasi'));
+        })->when($request->kbk, function($query) use ($request) {
+            $query->where('kbk', "{$request->kbk}");
+        })->when($request->kelompok, function($query) use ($request) {
+            $query->where('kelompok', "{$request->kelompok}");
+        })->paginate(16);
+        $pesertas->appends($request->only('search', 'delegasi', 'kbk', 'kelompok'));
 
         return view('peserta.main', compact('pesertas'))
             ->with('search', $request->search)
-            ->with('delegasi', $request->delegasi);
+            ->with('delegasi', $request->delegasi)
+            ->with('kelompok', $request->kelompok)
+            ->with('kbk', $request->kbk);
 
     }
 
@@ -93,8 +98,8 @@ class PesertaController extends Controller
             'nim' => $request->nim,
             'email' => $request->email,
             'alamat_kost' => $request->alamat_kost,
-            // 'kbk' => $request->kbk,
-            // 'kelompok' => $request->kelompok,
+            'kbk' => $request->kbk,
+            'kelompok' => $request->kelompok,
             'jurusan' => $request->jurusan,
             'delegasi' => $request->delegasi,
             'asal' => $request->asal,
@@ -151,28 +156,32 @@ class PesertaController extends Controller
 
         $peserta = Peserta::findOrFail($id);
         // Upload Image
-        if ($request->has('avatar')) {
+        if ($request->pict != $peserta->imageurl) {
+            if ($request->has('avatar')) {
 
-            $imagePath = 'images/peserta/'.$peserta->imageurl;
-            if ($peserta->imageurl != 'defaultimage.png') {
-                if (File::exists($imagePath)) {
-                    File::delete($imagePath);
+                $imagePath = 'images/peserta/'.$peserta->imageurl;
+                if ($peserta->imageurl != 'defaultimage.png') {
+                    if (File::exists($imagePath)) {
+                        File::delete($imagePath);
+                    }
                 }
+                $image = $request->avatar;
+
+                list($type, $image) = explode(';', $image);
+                list(, $image)      = explode(',', $image);
+
+                $image = base64_decode($image);
+                $imageName = time().'.png';
+                $path = public_path() . "/images/peserta/" . $imageName;
+                file_put_contents($path, $image);
+
+            } else {
+
+                $imageName = 'defaultimage.png';
+
             }
-            $image = $request->avatar;
-
-            list($type, $image) = explode(';', $image);
-            list(, $image)      = explode(',', $image);
-
-            $image = base64_decode($image);
-            $imageName = time().'.png';
-            $path = public_path() . "/images/peserta/" . $imageName;
-            file_put_contents($path, $image);
-
         } else {
-
-            $imageName = 'defaultimage.png';
-
+            $imageName = $peserta->imageurl;
         }
 
         $peserta->update([
@@ -182,8 +191,8 @@ class PesertaController extends Controller
             'nim' => $request->nim,
             'email' => $request->email,
             'alamat_kost' => $request->alamat_kost,
-            // 'kbk' => $request->kbk,
-            // 'kelompok' => $request->kelompok,
+            'kbk' => $request->kbk,
+            'kelompok' => $request->kelompok,
             'panggilan' => $request->panggilan,
             'jurusan' => $request->jurusan,
             'delegasi' => $request->delegasi,
@@ -237,7 +246,8 @@ class PesertaController extends Controller
         $pesertas = Peserta::when($request->delegasi, function($query) use ($request) {
             $query->where('delegasi', "{$request->delegasi}");
         })->get();
-    	return view('peserta.calendar', compact('pesertas'));
+    	return view('peserta.calendar', compact('pesertas'))
+            ->with('delegasi', $request->delegasi);
     }
 
 }
